# packages/contracts/tsp-output/openapi/openapi.yaml
openapi: 3.0.0
info:
  title: (title)
  description: |-
    Project API - RESTful API following modern best practices

    Key Features:
    - URL versioning (/v1)
    - Multi-tenant architecture (path-based with /orgs/{orgId})
    - Comprehensive pagination with navigation links
    - Structured error responses with detailed validation
    - Soft delete support with restore capability
    - Batch operations for bulk processing
    - Field selection and relation inclusion
    - Full-text search and filtering
  version: 0.0.0
...
paths:
...
  /v1/notifications:
    get:
      operationId: Notifications_list
      summary: List notifications
      description: |-
        List user's notification history
        Returns notifications for the authenticated user, excluding soft-deleted ones
      parameters:
        - name: page
          in: query
          required: false
          description: 'Page number (1-indexed, default: 1)'
          schema:
            type: integer
            format: int32
            default: 1
          explode: false
        - name: pageSize
          in: query
          required: false
          description: 'Number of items per page (max: 100, default: 20)'
          schema:
            type: integer
            format: int32
            default: 20
          explode: false
        - name: channel
          in: query
          required: false
          description: Filter by notification channel
          schema:
            $ref: '#/components/schemas/NotificationChannel'
          explode: false
        - name: category
          in: query
          required: false
          description: Filter by notification category
          schema:
            $ref: '#/components/schemas/NotificationCategory'
          explode: false
        - name: status
          in: query
          required: false
          description: Filter by notification status
          schema:
            $ref: '#/components/schemas/NotificationStatus'
          explode: false
      responses:
        '200':
          description: The request has succeeded.
          content:
            application/json:
              schema:
                anyOf:
                  - $ref: '#/components/schemas/NotificationListResponse'
                  - $ref: '#/components/schemas/ErrorResponse'
      tags:
        - Notifications
...
  /v1/notifications/send:
    post:
      operationId: Notifications_send
      summary: Send notification
      description: |-
        Send a notification
        Queues the notification for delivery through the specified channel
        Requires authentication. Consider adding authorization for production use.
      parameters: []
      responses:
        '200':
          description: The request has succeeded.
          content:
            application/json:
              schema:
                anyOf:
                  - $ref: '#/components/schemas/SendNotificationResponse'
                  - $ref: '#/components/schemas/ErrorResponse'
      tags:
        - Notifications
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SendNotificationRequest'
...
  /v1/notifications/{id}:
    get:
      operationId: Notifications_get
      summary: Get notification by ID
      description: |-
        Get a specific notification by ID
        Only returns notifications owned by the authenticated user
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: The request has succeeded.
          content:
            application/json:
              schema:
                anyOf:
                  - $ref: '#/components/schemas/NotificationResponse'
                  - $ref: '#/components/schemas/ErrorResponse'
      tags:
        - Notifications
    delete:
      operationId: Notifications_delete
      summary: Delete notification
      description: |-
        Soft delete a notification
        Marks the notification as deleted without actually removing it
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: The request has succeeded.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '204':
          description: 'There is no content to send for this request, but the headers may be useful. '
      tags:
        - Notifications
  /v1/notifications/{id}/read:
    patch:
      operationId: Notifications_markRead
      summary: Mark notification as read
      description: |-
        Mark a notification as read
        Updates the readAt timestamp for the notification
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: The request has succeeded.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '204':
          description: 'There is no content to send for this request, but the headers may be useful. '
      tags:
        - Notifications
  /v1/notifications/{id}/restore:
    post:
      operationId: Notifications_restore
      summary: Restore notification
      description: |-
        Restore a soft-deleted notification
        Clears the deletedAt timestamp to restore the notification
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: The request has succeeded.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '204':
          description: 'There is no content to send for this request, but the headers may be useful. '
      tags:
        - Notifications
  /v1/notifications/{id}/unread:
    patch:
      operationId: Notifications_markUnread
      summary: Mark notification as unread
      description: |-
        Mark a notification as unread
        Clears the readAt timestamp for the notification
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: The request has succeeded.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '204':
          description: 'There is no content to send for this request, but the headers may be useful. '
      tags:
        - Notifications
...
components:
  schemas:
...
    Error:
      type: object
      required:
        - code
        - message
        - requestId
      properties:
        code:
          type: string
          description: Machine-readable error code
        message:
          type: string
          description: Human-readable error message
        details:
          type: array
          items:
            $ref: '#/components/schemas/ErrorDetail'
          description: Array of detailed error information (e.g., validation errors)
        requestId:
          type: string
          description: Request ID for tracking and debugging
        documentationUrl:
          type: string
          description: URL to documentation about this error (optional)
      description: Standard error object
    ErrorDetail:
      type: object
      required:
        - field
        - code
        - message
      properties:
        field:
          type: string
          description: The field that caused the error (for validation errors)
        code:
          type: string
          description: Specific error code for this detail
        message:
          type: string
          description: Human-readable error message
        metadata:
          type: object
          additionalProperties: {}
          description: Additional metadata about the error
      description: Detailed error information for a specific field or validation
    ErrorResponse:
      type: object
      required:
        - error
      properties:
        error:
          allOf:
            - $ref: '#/components/schemas/Error'
          description: The error object
      description: Standard error response envelope
...
    Notification:
      type: object
      required:
        - id
        - channel
        - category
        - priority
        - status
        - retryCount
        - maxRetries
        - createdAt
        - updatedAt
      properties:
        id:
          type: string
          description: Unique notification identifier
          example: notif_abc123
        userId:
          type: string
          description: User ID who should receive this notification
          example: usr_xyz789
        channel:
          allOf:
            - $ref: '#/components/schemas/NotificationChannel'
          description: Notification channel
        category:
          allOf:
            - $ref: '#/components/schemas/NotificationCategory'
          description: Notification category
        priority:
          allOf:
            - $ref: '#/components/schemas/NotificationPriority'
          description: Notification priority
        status:
          allOf:
            - $ref: '#/components/schemas/NotificationStatus'
          description: Current notification status
        recipientEmail:
          type: string
          description: Email address this was sent to
        recipientPhone:
          type: string
          description: Phone number this was sent to
        recipientTelegramId:
          type: string
          description: Telegram ID this was sent to
        templateId:
          type: string
          description: Template ID used (if any)
        subject:
          type: string
          description: Subject line (for email)
        body:
          type: string
          description: Plain text body
        bodyHtml:
          type: string
          description: HTML body (for email)
        templateData:
          type: object
          additionalProperties: {}
          description: Template data used for rendering
        campaignId:
          type: string
          description: Campaign ID for bulk sends
        provider:
          type: string
          description: Provider used to send (e.g., "nodemailer", "twilio")
        providerMessageId:
          type: string
          description: Provider's message ID
        sentAt:
          type: string
          format: date-time
          description: Timestamp when notification was sent
        deliveredAt:
          type: string
          format: date-time
          description: Timestamp when notification was delivered
        failedAt:
          type: string
          format: date-time
          description: Timestamp when notification failed
        readAt:
          type: string
          format: date-time
          description: Timestamp when notification was read by user
        deletedAt:
          type: string
          format: date-time
          description: Timestamp when notification was soft deleted
        retryCount:
          type: integer
          format: int32
          description: Number of retry attempts
        maxRetries:
          type: integer
          format: int32
          description: Maximum retry attempts allowed
        nextRetryAt:
          type: string
          format: date-time
          description: Next retry timestamp
        statusMessage:
          type: string
          description: Status message (error details if failed)
        metadata:
          type: object
          additionalProperties: {}
          description: Metadata for custom data
        createdAt:
          type: string
          format: date-time
          description: Timestamp when notification was created
        updatedAt:
          type: string
          format: date-time
          description: Timestamp when notification was last updated
      description: Notification resource model
    NotificationCategory:
      type: string
      enum:
        - transactional
        - marketing
        - security
        - system
      description: Notification category enum
    NotificationChannel:
      type: string
      enum:
        - email
        - sms
        - whatsapp
        - telegram
        - push
      description: Notification channel enum
    NotificationListResponse:
      type: object
      required:
        - data
        - pagination
        - meta
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/Notification'
        pagination:
          $ref: '#/components/schemas/Pagination'
        meta:
          $ref: '#/components/schemas/ResponseMeta'
      description: Notification collection response
...
    NotificationPriority:
      type: string
      enum:
        - urgent
        - high
        - normal
        - low
      description: Notification priority enum
    NotificationRecipient:
      type: object
      properties:
        email:
          type: string
          description: Email address (required for email channel)
          example: user@example.com
        phone:
          type: string
          description: Phone number in E.164 format (required for SMS/WhatsApp)
          example: '+1234567890'
        telegramId:
          type: string
          description: Telegram chat ID (required for Telegram channel)
          example: '123456789'
      description: Notification recipient information
    NotificationResponse:
      type: object
      required:
        - data
        - meta
      properties:
        data:
          $ref: '#/components/schemas/Notification'
        meta:
          $ref: '#/components/schemas/ResponseMeta'
      description: Single notification response
    NotificationStatus:
      type: string
      enum:
        - pending
        - queued
        - processing
        - sent
        - delivered
        - failed
        - bounced
      description: Notification status enum
    Pagination:
      type: object
      required:
        - page
        - pageSize
        - totalPages
        - totalCount
        - hasNext
        - hasPrevious
      properties:
        page:
          type: integer
          format: int32
          description: Current page number (1-indexed)
        pageSize:
          type: integer
          format: int32
          description: Number of items per page
        totalPages:
          type: integer
          format: int32
          description: Total number of pages
        totalCount:
          type: integer
          format: int32
          description: Total number of items across all pages
        hasNext:
          type: boolean
          description: Whether there is a next page
        hasPrevious:
          type: boolean
          description: Whether there is a previous page
        links:
          allOf:
            - $ref: '#/components/schemas/PaginationLinks'
          description: Navigation links for pagination
        requestedPageSize:
          type: integer
          format: int32
          description: Original requested page size (included if pageSize was capped)
      description: Page-based pagination metadata for collection responses
    PaginationLinks:
      type: object
      required:
        - first
        - previous
        - next
        - last
      properties:
        first:
          type: string
          nullable: true
          description: URL for the first page
        previous:
          type: string
          nullable: true
          description: URL for the previous page (null if on first page)
        next:
          type: string
          nullable: true
          description: URL for the next page (null if on last page)
        last:
          type: string
          nullable: true
          description: URL for the last page
      description: Pagination links for navigating through paginated results
...
    ResponseMeta:
      type: object
      required:
        - requestId
        - timestamp
      properties:
        requestId:
          type: string
          description: Unique request identifier for tracking and debugging
        timestamp:
          type: string
          format: date-time
          description: Timestamp when the response was generated
        apiVersion:
          type: string
          description: API version used for this request
        tenantId:
          type: string
          description: Tenant/organization ID for multi-tenant requests
        tenantName:
          type: string
          description: Tenant/organization name for multi-tenant requests
        durationMs:
          type: integer
          format: int32
          description: Processing time in milliseconds
      description: Response metadata included in all API responses
...
    SendNotificationRequest:
      type: object
      required:
        - channel
        - category
        - recipient
      properties:
        channel:
          allOf:
            - $ref: '#/components/schemas/NotificationChannel'
          description: Notification channel
        category:
          allOf:
            - $ref: '#/components/schemas/NotificationCategory'
          description: Notification category
        priority:
          allOf:
            - $ref: '#/components/schemas/NotificationPriority'
          description: 'Notification priority (default: normal)'
        recipient:
          allOf:
            - $ref: '#/components/schemas/NotificationRecipient'
          description: Recipient information
        templateId:
          type: string
          description: Template ID to use for rendering
          example: welcome-email
        templateData:
          type: object
          additionalProperties: {}
          description: Data for template rendering
        subject:
          type: string
          description: Subject line (for email, required if not using template)
          example: Welcome to our platform
        body:
          type: string
          description: Plain text body (required if not using template)
          example: Thank you for signing up!
        bodyHtml:
          type: string
          description: HTML body (for email)
        metadata:
          type: object
          additionalProperties: {}
          description: Custom metadata
      description: Request body for sending a notification
    SendNotificationResponse:
      type: object
      required:
        - data
        - meta
      properties:
        data:
          type: object
          properties:
            messageId:
              type: string
              description: Message ID assigned to the notification
            provider:
              type: string
              description: Provider used to send
          required:
            - messageId
            - provider
        meta:
          $ref: '#/components/schemas/ResponseMeta'
      description: Send notification success response
...
servers:
  - url: https://api.example.com
    description: Production server
    variables: {}
  - url: http://localhost:3001
    description: Development server
    variables: {}